requirements.txt
fastapi==0.112.1
motor==3.5.1
pydantic-settings==2.4.0
python-dotenv==1.0.1
